# hasconversion for tensorspace