--- a/src/Spaces/PolynomialSpace.jl
+++ b/src/Spaces/PolynomialSpace.jl
@@ -412,8 +412,14 @@ function Fun(::typeof(identity), S::NormalizedPolynomialSpace)
     Fun(S, coeffs)
 end
 
+function _conversion_maxspace_rule(a::NormalizedPolynomialSpace{S}, b::S) where S<:PolynomialSpace
+    domainscompatible(domain(a), domain(b)) ? b : NoSpace()
+end
 function conversion_rule(a::NormalizedPolynomialSpace{S}, b::S) where S<:PolynomialSpace
-    domainscompatible(domain(a), domain(b)) ? a : NoSpace()
+    _conversion_maxspace_rule(a, b)
+end
+function maxspace_rule(a::NormalizedPolynomialSpace{S}, b::S) where S<:PolynomialSpace
+    _conversion_maxspace_rule(a, b)
 end
 
 bandwidths(C::ConcreteConversion{NormalizedPolynomialSpace{S,D,R},S}) where {S,D,R} = (0, 0)
@@ -455,6 +461,25 @@ spacescompatible(a::NormalizedPolynomialSpace,b::NormalizedPolynomialSpace) = sp
 hasconversion(a::PolynomialSpace,b::NormalizedPolynomialSpace) = hasconversion(a,b.space)
 hasconversion(a::NormalizedPolynomialSpace,b::PolynomialSpace) = hasconversion(a.space,b)
 hasconversion(a::NormalizedPolynomialSpace,b::NormalizedPolynomialSpace) = hasconversion(a.space,b)
+
+# Tensor products of normalized and unnormalized spaces may have banded conversions defined
+# A banded conversion exists in special cases, where both conversion operators are diagonal
+_stripnorm(N::NormalizedPolynomialSpace) = canonicalspace(N)
+_stripnorm(x::PolynomialSpace) = x
+function _hasconversion_tensor(A, B)
+    A1, A2 = A
+    B1, B2 = B
+
+    _stripnorm(A1) == _stripnorm(B1) && _stripnorm(A2) == _stripnorm(B2)
+end
+const MaybeNormalized{S<:PolynomialSpace} = Union{S, NormalizedPolynomialSpace{S}}
+const MaybeNormalizedTensorSpace{P1,P2} = TensorSpace{<:Tuple{MaybeNormalized{P1},MaybeNormalized{P2}}}
+
+function hasconversion(A::MaybeNormalizedTensorSpace{<:P1, <:P2},
+        B::MaybeNormalizedTensorSpace{<:P1, <:P2}) where {P1<:PolynomialSpace,P2<:PolynomialSpace}
+
+    _hasconversion_tensor(factors(A), factors(B))
+end
 
 
 function Multiplication(f::Fun{<:PolynomialSpace}, sp::NormalizedPolynomialSpace)