After this, the following works:
```julia
julia> ApproxFunBase.hasconversion(Chebyshev()*Legendre(), Chebyshev()*NormalizedLegendre())
true
```
On master
```julia
julia> f = Fun((x,y)->x^10*y^4 + x^2*y^3, Chebyshev()*Legendre());
```
TTFX
```julia
julia> @time Fun(f, Chebyshev()*NormalizedLegendre());
 12.396606 seconds (41.16 M allocations: 2.226 GiB, 6.72% gc time, 99.99% compilation time)
```
Benchmark
```julia
julia> @btime Fun($f, Chebyshev()*NormalizedLegendre());
  287.812 μs (416 allocations: 66.70 KiB)
```
This PR
TTFX
```julia
julia> @time Fun(f, Chebyshev()*NormalizedLegendre());
 10.915021 seconds (38.30 M allocations: 2.089 GiB, 6.96% gc time, 99.99% compilation time)
```
Benchmark
```julia
julia> @btime Fun($f, Chebyshev()*NormalizedLegendre());
  66.401 μs (1386 allocations: 113.58 KiB)
```
The allocations go up, but the execution time is lower.